my_project/requirements.txt
flask
fastapi
uvicorn
psycopg2
celery
redis
requests
python-dotenv
openai
supabase
gunicorn